refactor: enhance confirmation handling in CLI commands
- Updated `require_confirmation` function to include optional summary, notes, and warnings for better context in non-interactive modes.
- Improved error handling for non-interactive sessions, emitting structured errors in JSON format.
- Refactored `eval_submit` and `submit` commands to utilize the updated confirmation logic.
- Added `print_remote_fetch_notice` utility to inform users about remote code fetching and local config usage.
- Adjusted `PaginatedCloudEvalRuns` to expose `eval_runs` for consistency with other paginated models.
