+++ osmosis_ai/cli/commands/train.py
@@ -21,178 +21,6 @@ def _detail_fields(rows: list[tuple[str, str]]) -> list[Any]:
     return [DetailField(label=label, value=value) for label, value in rows]
 
 
-def _require_confirmation(
-    message: str,
-    *,
-    yes: bool,
-    summary: list[tuple[str, str]] | None = None,
-    notes: list[str] | None = None,
-    warnings: list[str] | None = None,
-) -> None:
-    """Confirm a destructive action, surfacing context in non-interactive modes.
-
-    In rich+interactive mode, falls back to the questionary prompt. In JSON
-    and plain modes (or any non-interactive shell), raises an
-    ``INTERACTIVE_REQUIRED`` error that carries the prompt question, the
-    summary of what would be acted on, and any notes/warnings — so AI
-    agents and CI scripts can see exactly what they're being asked to
-    confirm without first having to retry with ``--yes``.
-    """
-    if yes:
-        return
-
-    from osmosis_ai.cli.output import OutputFormat, get_output_context
-
-    output = get_output_context()
-    if output.format is not OutputFormat.rich or not output.interactive:
-        details: dict[str, Any] = {"prompt": message}
-        if summary:
-            details["summary"] = {label: value for label, value in summary}
-        if notes:
-            details["notes"] = list(notes)
-        if warnings:
-            details["warnings"] = list(warnings)
-
-        if output.format is OutputFormat.plain:
-            import sys
-
-            lines: list[str] = [f"Confirmation required: {message}"]
-            if summary:
-                for label, value in summary:
-                    lines.append(f"  {label}: {value}")
-            if notes:
-                lines.append("Notes:")
-                for note in notes:
-                    lines.append(f"  - {note}")
-            if warnings:
-                lines.append("Warnings:")
-                for warning in warnings:
-                    lines.append(f"  - {warning}")
-            sys.stderr.write("\n".join(lines) + "\n")
-            sys.stderr.flush()
-
-        err = CLIError(
-            "Use --yes to confirm in non-interactive mode.",
-            code="INTERACTIVE_REQUIRED",
-            details=details,
-        )
-        if output.format is OutputFormat.json:
-            from osmosis_ai.cli.output import emit_structured_error_to_stderr
-
-            emit_structured_error_to_stderr(err)
-            raise typer.Exit(1)
-        raise err
-
-    from osmosis_ai.cli.prompts import require_confirmation
-
-    require_confirmation(message, yes=yes)
-
-
-def _print_remote_fetch_notice(
-    workspace_directory: Path,
-    *,
-    pinned_commit_sha: str | None,
-) -> tuple[list[str], list[str]]:
-    """Remind the user that submit pulls *code* from the connected Git remote
-    while reading *config values* from the local TOML file.
-
-    The platform resolves training code from the Platform-connected
-    repository (or fetches a pinned commit) before training, so local
-    *code* changes that haven't been pushed will silently be ignored.
-    The config TOML passed to ``osmosis train submit``, by contrast, is
-    read from disk and its values are sent verbatim in the submit
-    payload — local edits to the config take effect immediately, even
-    if they are uncommitted.
-
-    Returns ``(notes, warnings)`` as plain-text lists so callers can
-    surface the same context in non-rich modes (e.g. the JSON error
-    envelope when ``--yes`` is missing). The Rich panel is rendered
-    only when the output format is Rich.
-    """
-    from osmosis_ai.cli.output import OutputFormat, get_output_context
-    from osmosis_ai.platform.cli.workspace_repo import summarize_local_git_state
-
-    state = summarize_local_git_state(workspace_directory)
-
-    warnings: list[str] = []
-    if state is not None:
-        if state.is_dirty:
-            warnings.append(
-                "Uncommitted changes detected — code edits won't be picked up "
-                "(only the config file above is read locally)."
-            )
-        if state.has_upstream and state.ahead > 0:
-            commits_word = "commit" if state.ahead == 1 else "commits"
-            warnings.append(
-                f"{state.ahead} unpushed {commits_word} ahead of upstream — "
-                "push code before submitting."
-            )
-        elif state.branch is not None and not state.has_upstream:
-            warnings.append(
-                f"Branch '{state.branch}' has no upstream — "
-                "push code and set tracking before submitting."
-            )
-
-    notes: list[str] = []
-    if pinned_commit_sha:
-        notes.append(
-            f"Osmosis will fetch commit {pinned_commit_sha} from the "
-            "Platform-connected repository for training code."
-        )
-        notes.append("Make sure that commit is pushed to origin.")
-    else:
-        notes.append(
-            "Osmosis will fetch training code from the Platform-connected repository."
-        )
-        if state is not None and state.branch and state.head_sha:
-            notes.append(f"Local branch: {state.branch} @ {state.head_sha[:8]}")
-        notes.append("Make sure your code changes are committed and pushed.")
-        warnings.append(
-            "Platform source selection may differ from your local branch when no "
-            "commit_sha is set."
-        )
-    notes.append(
-        "Config values come from your local TOML file and are submitted "
-        "as-is — uncommitted edits to the config still apply."
-    )
-
-    if get_output_context().format is OutputFormat.rich:
-        body_lines: list[str] = []
-        if pinned_commit_sha:
-            body_lines.append(
-                f"Osmosis will fetch commit [bold]{console.escape(pinned_commit_sha)}[/bold] "
-                "from the Platform-connected repository for training code."
-            )
-            body_lines.append("Make sure that commit is pushed to origin.")
-        else:
-            body_lines.append(
-                "Osmosis will fetch training code from the Platform-connected repository."
-            )
-            if state is not None and state.branch and state.head_sha:
-                body_lines.append(
-                    f"Local: [bold]{console.escape(state.branch)}[/bold] @ "
-                    f"[dim]{console.escape(state.head_sha[:8])}[/dim]"
-                )
-            body_lines.append("Make sure your code changes are committed and pushed.")
-
-        body_lines.append("")
-        body_lines.append(
-            "[dim]Config values above come from your local TOML file and are "
-            "submitted as-is — uncommitted edits to the config still apply.[/dim]"
-        )
-
-        if warnings:
-            body_lines.append("")
-            for warning in warnings:
-                body_lines.append(f"[yellow]• {console.escape(warning)}[/yellow]")
-
-        style = "yellow" if warnings else "blue"
-        title = "Push before submitting" if warnings else "Before you submit"
-        console.panel(title, "\n".join(body_lines), style=style)
-
-    return notes, warnings
-
-
 @app.command("list")
 def list_runs(
     limit: int = typer.Option(
@@ -516,11 +344,15 @@ def submit(
 ) -> Any:
     """Submit a new training run."""
     from osmosis_ai.cli.output import OperationResult, get_output_context
+    from osmosis_ai.cli.prompts import require_confirmation
     from osmosis_ai.platform.cli.training_config import (
         load_training_config,
         validate_training_context_paths,
     )
-    from osmosis_ai.platform.cli.utils import require_git_workspace_directory_context
+    from osmosis_ai.platform.cli.utils import (
+        print_remote_fetch_notice,
+        require_git_workspace_directory_context,
+    )
     from osmosis_ai.platform.cli.workspace_directory_context import git_result_context
     from osmosis_ai.platform.cli.workspace_directory_contract import (
         ensure_workspace_directory_config_path,
@@ -581,12 +413,12 @@ def submit(
         title="Training Run",
     )
 
-    notes, warnings = _print_remote_fetch_notice(
+    notes, warnings = print_remote_fetch_notice(
         workspace_directory,
         pinned_commit_sha=config.experiment_commit_sha,
     )
 
-    _require_confirmation(
+    require_confirmation(
         "Submit this training run?",
         yes=yes,
         summary=summary_rows,
@@ -714,6 +546,7 @@ def stop(
 ) -> Any:
     """Stop a training run."""
     from osmosis_ai.cli.output import OperationResult, get_output_context
+    from osmosis_ai.cli.prompts import require_confirmation
     from osmosis_ai.platform.cli.utils import require_git_workspace_directory_context
     from osmosis_ai.platform.cli.workspace_directory_context import git_result_context
 
@@ -724,7 +557,7 @@ def stop(
 
     client = OsmosisClient()
 
-    _require_confirmation(
+    require_confirmation(
         f'Stop training run "{name}"?',
         yes=yes,
         summary=[("Name", name)],
